# api/openapi.yaml
openapi: 3.1.0
info:
  title: Marketing Value Creator (MVC) API
  version: "1.0.0"
  description: |
    Enterprise generative AI campaign planning platform for Nova Electronics Corp.
    Orchestrates 4 specialized sub-agents ([P1] Market Sensing, [P2] Strategy & Brief,
    [P3] Creative Content, [P4] Performance & Insights) with Human-in-the-Loop (HITL) review gates.
servers:
  - url: /
    description: Current host
security:
  - cookieAuth: []
  - googleOAuth: []
paths:
...
  /api/v1/campaigns/parse-prompt:
    post:
      summary: Parse natural language prompt into structured campaign brief parameters
      operationId: parseCampaignPrompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParsePromptRequest'
      responses:
        "200":
          description: Successfully parsed parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParsePromptResponse'
        "400":
          description: Model Armor rejection or invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/campaigns:
    get:
      summary: List campaigns owned by authenticated user
      operationId: listUserCampaigns
      responses:
        "200":
          description: List of campaigns owned by the current user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CampaignSessionResponse'
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Start a new multi-agent campaign planning DAG
      operationId: createCampaign
      description: Initializes a campaign workflow and initiates P1 (Market Sensing). Returns session details or streams SSE events.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
      responses:
        "200":
          description: Campaign initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignSessionResponse'
        "400":
          description: Invalid request or Model Armor prompt rejection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized / invalid Google OAuth ID token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/campaigns/{sessionId}/rollback:
    post:
      summary: Rollback campaign session strictly to immediately preceding stage (N - 1)
      operationId: rollbackCampaignStage
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session state successfully rolled back to preceding stage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignSessionResponse'
        "400":
          description: Invalid rollback attempt (cannot rollback from initial stage or skip stages)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized / invalid Google OAuth ID token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Campaign session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/campaigns/{sessionId}:
    get:
      summary: Get current campaign session state and deliverables
      operationId: getCampaignSession
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session state details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignSessionResponse'
        "404":
          description: Campaign session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/campaigns/{sessionId}/visual:
    get:
      summary: Access campaign visual via 307 redirect to V4 Signed URL or in-memory draft
      operationId: getCampaignVisual
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        "307":
          description: Temporary redirect to Google Cloud Storage V4 Signed URL
          headers:
            Location:
              schema:
                type: string
                description: V4 Signed URL on storage.googleapis.com
        "200":
          description: Image binary stream (for in-memory draft or streaming fallback)
          content:
            image/png:
              schema:
                type: string
                format: binary
        "404":
          description: Visual asset or campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/campaigns/{sessionId}/draft-image:
    get:
      summary: Fetch in-memory draft marketing visual before approval
      operationId: getCampaignDraftImage
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Image binary stream of in-memory draft visual
          content:
            image/png:
              schema:
                type: string
                format: binary
        "404":
          description: No in-memory draft image found for session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/campaigns/{sessionId}/visual-token:
    get:
      summary: Get ephemeral V4 Signed URL token for campaign visual
      operationId: getCampaignVisualToken
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ephemeral V4 Signed URL token and expiration
          content:
            application/json:
              schema:
                type: object
                properties:
                  signedUrl:
                    type: string
                    example: "https://storage.googleapis.com/bucket/path?X-Goog-Signature=..."
                  expiresIn:
                    type: integer
                    example: 3600
        "404":
          description: Visual asset or campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/campaigns/{sessionId}/approve:
    post:
      summary: Submit human-in-the-loop review approval or revision feedback
      operationId: approveOrReviseCampaignStage
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StageApprovalRequest'
      responses:
        "200":
          description: Stage transition processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignSessionResponse'
        "400":
          description: Invalid stage state or feedback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mvc_session
      description: Session cookie stored in Cloud SQL
    googleOAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Google OAuth 2.0 OIDC ID Token
  schemas:
    CreateCampaignRequest:
      type: object
      required:
        - brandName
        - productName
        - campaignObjective
        - targetAudience
        - budgetAmount
        - currency
      properties:
        brandName:
          type: string
          example: "Nova Electronics Corp"
        productName:
          type: string
          example: "Galaxy S27 Ultra"
        campaignObjective:
          type: string
          example: "Black Friday Global Campaign targeting premium tech enthusiasts with AI camera features."
        targetAudience:
          type: string
          example: "Tech-savvy professionals and mobile photographers aged 25-45."
        budgetAmount:
          type: number
          example: 1000000.0
        currency:
          type: string
          default: "USD"
          example: "USD"
        language:
          type: string
          enum: ["ko", "en"]
          default: "ko"
          example: "ko"
        channels:
          type: array
          items:
            type: string
          example: ["Digital Video", "Paid Search", "Social Media", "Retail Display"]
        stream:
          type: boolean
          default: true
          description: Whether to return response as Server-Sent Events (SSE)
    ParsePromptRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
          example: "20대 타겟 브랜드 인지도 캠페인 기획해줘"
        language:
          type: string
          enum: ["ko", "en"]
          default: "ko"
    ParsePromptResponse:
      type: object
      properties:
        brandName:
          type: string
          example: ""
        productName:
          type: string
          example: ""
        campaignObjective:
          type: string
          example: "브랜드 인지도 캠페인 기획"
        targetAudience:
          type: string
          example: "20대"
        budgetAmount:
          type: number
          nullable: true
          example: null
        currency:
          type: string
          default: "USD"
          example: "KRW"
        channels:
          type: array
          items:
            type: string
          example: []
    StageApprovalRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: ["approve", "revise"]
          description: "'approve' advances to next stage; 'revise' re-executes current stage with feedback"
        feedback:
          type: string
          description: Marketer text feedback for revision
          example: "Please emphasize trade-in discounts more prominently in the brief and visual prompt."
        stream:
          type: boolean
          default: true
        deliverableUpdates:
          type: object
          additionalProperties: true
          description: "Optional dictionary containing updated deliverable fields modified by marketer"
    CampaignSessionResponse:
      type: object
      required:
        - sessionId
        - status
        - currentStage
        - brandName
        - productName
        - createdAt
        - updatedAt
      properties:
        sessionId:
          type: string
          example: "camp-9f8a3c21"
        userId:
          type: string
          nullable: true
          example: "a8098c1a-f86e-11da-bd1a-00112444be1e"
        status:
          type: string
          enum: ["INITIALIZING", "RUNNING", "PAUSED_FOR_REVIEW", "COMPLETED", "FAILED"]
          example: "PAUSED_FOR_REVIEW"
        currentStage:
          type: string
          enum: ["MARKET_SENSING", "STRATEGY_BRIEF", "CREATIVE_CONTENT", "PERFORMANCE_INSIGHTS", "MEDIA_EXECUTION", "COMPLETED"]
          example: "MARKET_SENSING"
        tenantId:
          type: string
          example: "default"
        channels:
          type: array
          items:
            type: string
        revisionCount:
          type: integer
          example: 0
        brandName:
          type: string
        productName:
          type: string
        campaignObjective:
          type: string
        budgetAmount:
          type: number
        currency:
          type: string
        deliverables:
          type: object
          properties:
            marketSensing:
              $ref: '#/components/schemas/MarketSensingDeliverable'
            campaignBrief:
              $ref: '#/components/schemas/CampaignBriefDeliverable'
            creativeContent:
              $ref: '#/components/schemas/CreativeContentDeliverable'
            performanceInsights:
              $ref: '#/components/schemas/PerformanceInsightsDeliverable'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MarketSensingDeliverable:
      type: object
      required:
        - targetMarket
        - consumerTrends
        - competitiveAnalysis
        - sentimentOverview
        - strategicOpportunities
      properties:
        targetMarket: { type: string }
        consumerTrends:
          type: array
          items: { type: string }
        competitiveAnalysis:
          type: array
          items:
            type: object
            required: [competitor, strengths, vulnerabilities]
            properties:
              competitor: { type: string }
              strengths: { type: array, items: { type: string } }
              vulnerabilities: { type: array, items: { type: string } }
        sentimentOverview:
          type: object
          required: [positiveThemes, frictionPoints, overallSentimentScore]
          properties:
            positiveThemes: { type: array, items: { type: string } }
            frictionPoints: { type: array, items: { type: string } }
            overallSentimentScore: { type: number }
        strategicOpportunities:
          type: array
          items: { type: string }
    CampaignBriefDeliverable:
      type: object
      required:
        - campaignTitle
        - coreValueProposition
        - targetPersonas
        - messagingPillars
        - toneAndVoice
      properties:
        campaignTitle: { type: string }
        coreValueProposition: { type: string }
        targetPersonas:
          type: array
          items:
            type: object
            required: [name, demographics, primaryNeeds, barriers]
            properties:
              name: { type: string }
              demographics: { type: string }
              primaryNeeds: { type: array, items: { type: string } }
              barriers: { type: array, items: { type: string } }
        messagingPillars:
          type: array
          items:
            type: object
            required: [pillar, keyMessage, proofPoints]
            properties:
              pillar: { type: string }
              keyMessage: { type: string }
              proofPoints: { type: array, items: { type: string } }
        toneAndVoice:
          type: array
          items: { type: string }
    CreativeContentDeliverable:
      type: object
      required:
        - visualConceptTitle
        - visualPromptUsed
        - assetUrl
        - headlineCopy
        - bodyCopy
        - callToAction
      properties:
        visualConceptTitle: { type: string }
        visualPromptUsed: { type: string }
        assetUrl:
          type: string
          description: GCS URI or HTTP URL of generated marketing image
        storageUri:
          type: string
          nullable: true
          description: Direct gs:// or https:// GCS storage URI of committed asset
        headlineCopy: { type: string }
        bodyCopy: { type: string }
        callToAction: { type: string }
        aspectRatio: { type: string, default: "16:9" }
        generationTimestamp: { type: string, format: date-time }
    PerformanceInsightsDeliverable:
      type: object
      required:
        - totalBudget
        - currency
        - channelAllocations
        - projectedKpis
        - expectedRoas
        - recommendations
      properties:
        totalBudget: { type: number }
        currency: { type: string }
        channelAllocations:
          type: array
          items:
            type: object
            required: [channel, allocationAmount, percentage, rationale]
            properties:
              channel: { type: string }
              allocationAmount: { type: number }
              percentage: { type: number }
              rationale: { type: string }
        projectedKpis:
          type: object
          required: [estimatedImpressions, estimatedClicks, estimatedConversions, projectedCtr]
          properties:
            estimatedImpressions: { type: integer }
            estimatedClicks: { type: integer }
            estimatedConversions: { type: integer }
            projectedCtr: { type: number }
        expectedRoas: { type: number }
        recommendations:
          type: array
          items: { type: string }
        creativeAssetUrl:
          type: string
          nullable: true
          description: "Visual asset URL evaluated for conversion and CTR impact"
        visualConceptSummary:
          type: string
          nullable: true
          description: "Creative concept summary evaluated during performance modeling"
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - traceId
      properties:
        error: { type: string }
        message: { type: string }
        traceId: { type: string }
        detail: { type: object }
...