# app/routers/campaigns.py
# Copyright 2026 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

"""Campaign lifecycle and multi-agent DAG orchestration routes."""

from typing import Any

from fastapi import APIRouter, Depends, HTTPException, status

from app.orchestrator.engine import (
    CampaignOrchestrationEngine,
    get_orchestration_engine,
)
from app.orchestrator.security import (
    SecurityManager,
    get_current_user,
    get_security_manager,
)
from app.orchestrator.session_repo import (
    SessionRepository,
    UserModel,
    get_session_repo,
)
from app.schemas.campaign import (
    CampaignSessionResponse,
    CreateCampaignRequest,
    ParsePromptRequest,
    ParsePromptResponse,
    StageApprovalRequest,
)
from app.schemas.errors import ErrorResponse

router = APIRouter(prefix="/api/v1/campaigns", tags=["Campaigns"])


@router.post(
    "/parse-prompt",
    response_model=ParsePromptResponse,
    responses={400: {"model": ErrorResponse}, 401: {"model": ErrorResponse}},
)
async def parse_campaign_prompt(
    payload: ParsePromptRequest,
    user: UserModel = Depends(get_current_user),
    security: SecurityManager = Depends(get_security_manager),
    engine: CampaignOrchestrationEngine = Depends(get_orchestration_engine),
) -> ParsePromptResponse:
    """Parse natural language prompt into structured campaign brief parameters."""
    await security.inspect_prompt_safety(payload.prompt)
    return await engine.parse_prompt(payload.prompt, language=payload.language)


@router.get(
    "",
    response_model=list[CampaignSessionResponse],
    responses={401: {"model": ErrorResponse}},
)
async def list_campaigns(
    limit: int = 50,
    user: UserModel = Depends(get_current_user),
    repo: SessionRepository = Depends(get_session_repo),
) -> list[CampaignSessionResponse]:
    """List recent campaigns belonging to authenticated user."""
    return await repo.list_user_campaigns(user.user_id, limit=limit)


@router.post(
    "",
    response_model=CampaignSessionResponse,
    responses={400: {"model": ErrorResponse}, 401: {"model": ErrorResponse}},
)
async def create_campaign(
    payload: CreateCampaignRequest,
    user: UserModel = Depends(get_current_user),
    security: SecurityManager = Depends(get_security_manager),
    engine: CampaignOrchestrationEngine = Depends(get_orchestration_engine),
) -> CampaignSessionResponse:
    """Start a new multi-agent campaign planning DAG."""
    await security.inspect_prompt_safety(payload.campaignObjective)

    return await engine.create_campaign(
        payload, principal=user.email, user_id=user.user_id
    )


@router.get(
    "/{sessionId}",
    response_model=CampaignSessionResponse,
    responses={401: {"model": ErrorResponse}, 404: {"model": ErrorResponse}},
)
async def get_campaign_session(
    sessionId: str,
    user: UserModel = Depends(get_current_user),
    repo: SessionRepository = Depends(get_session_repo),
) -> CampaignSessionResponse:
    """Retrieve full campaign session state and deliverables scoped by user."""
    session = await repo.get_session(sessionId, user_id=user.user_id)
    if not session:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=f"Campaign session '{sessionId}' not found.",
        )
    return session


@router.post(
    "/{sessionId}/approve",
    response_model=CampaignSessionResponse,
    responses={
        400: {"model": ErrorResponse},
        401: {"model": ErrorResponse},
        404: {"model": ErrorResponse},
    },
)
async def approve_stage(
    sessionId: str,
    payload: StageApprovalRequest,
    user: UserModel = Depends(get_current_user),
    security: SecurityManager = Depends(get_security_manager),
    engine: CampaignOrchestrationEngine = Depends(get_orchestration_engine),
) -> CampaignSessionResponse:
    """Submit human review approval or revision feedback."""
    if payload.feedback:
        await security.inspect_prompt_safety(payload.feedback)

    updated = await engine.approve_stage(
        sessionId, payload, principal=user.email, user_id=user.user_id
    )
    if not updated:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=f"Campaign session '{sessionId}' not found.",
        )
    return updated


@router.post(
    "/{sessionId}/rollback",
    response_model=CampaignSessionResponse,
    responses={
        400: {"model": ErrorResponse},
        401: {"model": ErrorResponse},
        404: {"model": ErrorResponse},
    },
)
async def rollback_stage(
    sessionId: str,
    user: UserModel = Depends(get_current_user),
    engine: CampaignOrchestrationEngine = Depends(get_orchestration_engine),
) -> CampaignSessionResponse:
    """Rollback session strictly to the immediately preceding stage (N - 1)."""
    return await engine.rollback_stage(sessionId, user_id=user.user_id)


@router.patch(
    "/{sessionId}",
    response_model=CampaignSessionResponse,
    responses={401: {"model": ErrorResponse}, 404: {"model": ErrorResponse}},
)
async def update_campaign_session(
    sessionId: str,
    payload: dict[str, Any],
    user: UserModel = Depends(get_current_user),
    repo: SessionRepository = Depends(get_session_repo),
) -> CampaignSessionResponse:
    """Update campaign session deliverables or fields directly."""
    session = await repo.get_session(sessionId, user_id=user.user_id)
    if not session:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=f"Campaign session '{sessionId}' not found.",
        )

    deliverables_update = payload.get("deliverables")
    current_delivs = session.deliverables.model_dump(mode="json")
    if deliverables_update and isinstance(deliverables_update, dict):
        for k, v in deliverables_update.items():
            if v is not None:
                if (
                    k in current_delivs
                    and isinstance(current_delivs[k], dict)
                    and isinstance(v, dict)
                ):
                    current_delivs[k].update(v)
                else:
                    current_delivs[k] = v

    updated = await repo.update_session(
        sessionId,
        deliverables=current_delivs,
        user_id=user.user_id,
    )
    if not updated:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=f"Campaign session '{sessionId}' could not be updated.",
        )
    return updated